The Course is Completed Web-App Created

+++ db/repository/blog.py
@@ -4,18 +4,20 @@ from db.models.blog import Blog
 
 def create_new_blog(blog: CreateBlog, db: Session, author_id: int):
    blog = Blog(
-       id=blog.id,
       title=blog.title,
       slug=blog.slug,
       content=blog.content,
-      author_id=author_id
+      author_id=author_id,
+      isActive=blog.isActive
    )
    db.add(blog)
    db.commit()
    db.refresh(blog)
    return blog
 
-
+def list_blogs(db: Session):
+    blogs = db.query(Blog).all()
+    return blogs
 
 def retrieve_blog(id: int , db: Session ) :
     blog =db.query(Blog).filter(Blog.id == id).first()
@@ -36,6 +38,8 @@ def delete_blog_by_ID(id: int, db: Session, author_id: int = 1):
     blog = db.query(Blog).filter(Blog.id == id).first()
     if not blog:
         return {"error ": f"Blog with ID {id} not found"}
+    if blog.author_id != author_id:
+        return {"error": "You do not have permission to delete this blog"}
     db.delete(blog)
     db.commit()
     return {"message": f"Blog with ID {id} deleted successfully"}
